test: epwait: fix race test hang by stopping after draining all data
On some system under tests, especially smaller VMs, the writer can finish
producing data, but the race subtest keeps waiting for more CQEs, so it
hangs on a final wait with nothing left to deliver.
Break the loop if drained >= expected where expected is LOOPS * NPIPES * 3
as the test writes "foo".

Signed-off-by: Sebastian Chlad <sebastian.chlad@suse.com>